Typo

+++ app/src/main/java/hu/bme/aut/szoftarch/kozkincsker/data/model/Feedback.kt
@@ -18,7 +18,7 @@ data class Feedback(
     var id: String = UUID.randomUUID().toString(),
     var mission: Mission = Mission(),
     var writer: User = User(),
-    var stars: Int = 0,
+    var stars: Double = 0.0,
     var comment: String = "",
     var commentState: CommentVisibility = CommentVisibility.SHOW
 ): Parcelable{
